machinelearningteam/summary-statistics-processor-service/cmd/service/requirements.txt
grpcio-tools
numpy

py -m grpc_tools.protoc -I.. --python_out=. --grpc_python_out=. ../../proto/api.proto